feat(): installed types, etc, tested connection
