+++ index.ts
@@ -1,6 +1,8 @@
-import { httpServer } from "./src/http_server/index.js";
+import { getWsServerStarted, httpServer } from "./src/http_server/index.js";
 
 const HTTP_PORT = 8181;
 
 console.log(`Start static http server on the ${HTTP_PORT} port!`);
 httpServer.listen(HTTP_PORT);
+
+getWsServerStarted();
